langchain_langgraph_projects/requirements.txt
langchain
langchain-core
langchain-google-genai
langgraph
langchain-chroma
pydantic
python-dotenv
duckduckgo-search
